MandatoryGroups UI

+++ mandatorymarkseditor.cpp
@@ -20,7 +20,7 @@ MandatoryMarksEditor::~MandatoryMarksEditor()
 
 void MandatoryMarksEditor::UpdateModels()
 {
-    MandatoryMarks_model.setQuery("SELECT * FROM \"mandatoryMarks\" ORDER BY \"accessLevel\" ASC;");
+    MandatoryMarks_model.setQuery("SELECT * FROM \"mandatoryMarks\" WHERE \"accessLevel\" >0 ORDER BY \"accessLevel\" ASC;");
     ui->MandatoryMarks_listView->setModel(&MandatoryMarks_model);
     ui->MandatoryMarks_listView->setModelColumn(1);
 
@@ -29,25 +29,39 @@ void MandatoryMarksEditor::UpdateModels()
 
 void MandatoryMarksEditor::on_addMandatoryMark_pushButton_clicked()
 {
- MandatoryMarkCreator *creator = new MandatoryMarkCreator(this);
- creator->setModal(true);
- creator->exec();
- delete creator;
+    MandatoryMarkCreator *creator = new MandatoryMarkCreator(this);
+    creator->setModal(true);
+    creator->exec();
+    delete creator;
 }
 
 void MandatoryMarksEditor::on_DeleteMandatoryMark_pushButton_clicked()
-{
+{ //Удаление мандатной метки
     QSqlQuery DeleteMandatoryMark;
-   const int currentId = MandatoryMarks_model.data(MandatoryMarks_model.index(ui->MandatoryMarks_listView->currentIndex().row(),0)).toInt();
-   const QString currentDescription = MandatoryMarks_model.data(MandatoryMarks_model.index(ui->MandatoryMarks_listView->currentIndex().row(),1)).toString();
-    if(!DeleteMandatoryMark.exec("DELETE FROM \"mandatoryMarks\" WHERE \"mandatoryMarkId\" = "+QString::number(currentId)+";"))
-       {
-        qDebug() << "Unable to delete Mandatory mark. SQL Error: " << DeleteMandatoryMark.lastError() << DeleteMandatoryMark.lastQuery();
+    const int currentId = MandatoryMarks_model.data(MandatoryMarks_model.index(ui->MandatoryMarks_listView->currentIndex().row(),0)).toInt();
+    const QString currentDescription = MandatoryMarks_model.data(MandatoryMarks_model.index(ui->MandatoryMarks_listView->currentIndex().row(),1)).toString();
+    if(!DeleteMandatoryMark.exec("SELECT * FROM rights WHERE \"mandatoryMark\" = "+QString::number(currentId)+";"))
+    {
+        qDebug() << "Unable to check mandatory groups. SQL Error: " << DeleteMandatoryMark.lastError() << DeleteMandatoryMark.lastQuery();
+        return;
+    }
+    DeleteMandatoryMark.first();
+    if(DeleteMandatoryMark.size()>0)
+    {
+        QMessageBox::critical(this, "Ошибка", "Мандатная метка просвоена " +QString::number(DeleteMandatoryMark.size())+" мандатным группам. \n Смените метку этих групп перед удалением.");
         return;
     }
     else
     {
-        QMessageBox::information(this, "Удаление мандатной метки", "Мандатная метка \""+currentDescription+"\" удалена");
-    }
 
+        if(!DeleteMandatoryMark.exec("DELETE FROM \"mandatoryMarks\" WHERE \"mandatoryMarkId\" = "+QString::number(currentId)+";"))
+        {
+            qDebug() << "Unable to delete Mandatory mark. SQL Error: " << DeleteMandatoryMark.lastError() << DeleteMandatoryMark.lastQuery();
+            return;
+        }
+        else
+        {
+            QMessageBox::information(this, "Удаление мандатной метки", "Мандатная метка \""+currentDescription+"\" удалена");
+        }
+    }
 }
